Replace all instances of getting type 'long' in code with 'long long', as 'long' doesn't seem to be a valid type for Clang++. With changes up to here, all targets build, except main.cpp and the main program binary.

+++ gadgets/cmr/CmrParametricMappingGadget.cpp
@@ -99,7 +99,7 @@ namespace Gadgetron {
 
         // -------------------------------------------------------------
 
-        size_t encoding = (size_t)std::get<long>(data->meta[0]["encoding"].front());
+        size_t encoding = (size_t)std::get<long long>(data->meta[0]["encoding"].front());
         GADGET_CHECK_RETURN(encoding<num_encoding_spaces_, GADGET_FAIL);
 
         std::string dataRole = std::get<std::string>(data->meta[0][GADGETRON_DATA_ROLE].front());
