+++ gadgets/mri_core/FloatToFixPointGadget.cpp
@@ -54,7 +54,7 @@ namespace Gadgetron
                     std::transform(image_in.data.begin(), image_in.data.end(), output_data.begin(), real_value);
 
                     if (image_out.meta.count(GADGETRON_IMAGE_WINDOWCENTER) && image_out.meta[GADGETRON_IMAGE_WINDOWCENTER].size() > 0) {
-                        long windowCenter = std::get<long>(image_out.meta[GADGETRON_IMAGE_WINDOWCENTER].front());
+                        long windowCenter = std::get<long long>(image_out.meta[GADGETRON_IMAGE_WINDOWCENTER].front());
                         image_out.meta[GADGETRON_IMAGE_WINDOWCENTER] = {windowCenter + self.intensity_offset};
                     }
                 }
