+++ gadgets/mri_core/generic_recon_gadgets/GenericReconPartialFourierHandlingGadget.cpp
@@ -55,7 +55,7 @@ namespace Gadgetron {
 
         // call the partial foureir
 
-        size_t encoding = (size_t)std::get<long>(recon_res.meta[0]["encoding"].front());
+        size_t encoding = (size_t)std::get<long long>(recon_res.meta[0]["encoding"].front());
         if (encoding > num_encoding_spaces) throw std::runtime_error("Illegal number of encoding spaces provided");
 
         // std::string dataRole = std::string(recon_res.meta[0].as_str(GADGETRON_DATA_ROLE));
@@ -77,9 +77,9 @@ namespace Gadgetron {
 
         if (recon_res.meta[0].count("sampling_limits_RO")) {
             auto& sl = recon_res.meta[0]["sampling_limits_RO"];
-            sampling_limits.kspace_encoding_step_0.minimum = (uint32_t)std::get<long>(sl[0]);
-            sampling_limits.kspace_encoding_step_0.center = (uint32_t)std::get<long>(sl[1]);
-            sampling_limits.kspace_encoding_step_0.maximum = (uint32_t)std::get<long>(sl[2]);
+            sampling_limits.kspace_encoding_step_0.minimum = (uint32_t)std::get<long long>(sl[0]);
+            sampling_limits.kspace_encoding_step_0.center = (uint32_t)std::get<long long>(sl[1]);
+            sampling_limits.kspace_encoding_step_0.maximum = (uint32_t)std::get<long long>(sl[2]);
         }
 
         if (!((sampling_limits.kspace_encoding_step_0.minimum >= 0) && (sampling_limits.kspace_encoding_step_0.maximum < RO)
@@ -91,9 +91,9 @@ namespace Gadgetron {
 
         if (recon_res.meta[0].count("sampling_limits_E1")) {
             auto& sl = recon_res.meta[0]["sampling_limits_E1"];
-            sampling_limits.kspace_encoding_step_1.minimum = (uint32_t)std::get<long>(sl[0]);
-            sampling_limits.kspace_encoding_step_1.center = (uint32_t)std::get<long>(sl[1]);
-            sampling_limits.kspace_encoding_step_1.maximum = (uint32_t)std::get<long>(sl[2]);
+            sampling_limits.kspace_encoding_step_1.minimum = (uint32_t)std::get<long long>(sl[0]);
+            sampling_limits.kspace_encoding_step_1.center = (uint32_t)std::get<long long>(sl[1]);
+            sampling_limits.kspace_encoding_step_1.maximum = (uint32_t)std::get<long long>(sl[2]);
         }
 
         if (!((sampling_limits.kspace_encoding_step_1.minimum >= 0) && (sampling_limits.kspace_encoding_step_1.maximum < E1)
@@ -105,9 +105,9 @@ namespace Gadgetron {
 
         if (recon_res.meta[0].count("sampling_limits_E2")) {
             auto& sl = recon_res.meta[0]["sampling_limits_E2"];
-            sampling_limits.kspace_encoding_step_2.minimum = (uint32_t)std::get<long>(sl[0]);
-            sampling_limits.kspace_encoding_step_2.center = (uint32_t)std::get<long>(sl[1]);
-            sampling_limits.kspace_encoding_step_2.maximum = (uint32_t)std::get<long>(sl[2]);
+            sampling_limits.kspace_encoding_step_2.minimum = (uint32_t)std::get<long long>(sl[0]);
+            sampling_limits.kspace_encoding_step_2.center = (uint32_t)std::get<long long>(sl[1]);
+            sampling_limits.kspace_encoding_step_2.maximum = (uint32_t)std::get<long long>(sl[2]);
         }
 
         if (!((sampling_limits.kspace_encoding_step_2.minimum >= 0) && (sampling_limits.kspace_encoding_step_2.maximum < E2)
